+++ src/core/game/game_timer.h
